Implement basic image selection
Wanted to dynamically import but found out that's not a thing? Well, it's being polyfilled, but seemed
like potentially more than I wanted to bite off at this exact moment in time. Simply hosting the images
elsewhere would be loads simpler but I feel like I know how to do that definitively, so why not mess
with some file loading?